Create Introduction for pandas for marketing
Data Science Techniques are used to understand the impact of marketing campaigns. Marketing campaigns mean anything that required the marketing team to put in work to promote your product. It could be new creative direction, a discounted product, targeting a specific demographic. 
What does a data scientist on a marketing team do ? 
Well possibilities are endless, but there are main tasks we should aware of them. 
1. Analyzing marketing campaign performance
2. Attributing credit for conversion to marketing channels. For ex: when you send out an email how many users subscribe ? Given current conversion rates and revenue, should you continue investing in this channel and how much should you spend?  
3. A/B testing (running experiments, to try to understand the impact of a particular change )